+++ changelog.md
@@ -3,6 +3,7 @@
 ### v1.5.0
 
 - updated readme w/ updated UI
+- application now supports audio preview while playing videos
 
 ### v1.4.3